Moved ConsensusMetrics out of EraSupervisor.
