+++ node/src/components/consensus.rs
@@ -3,6 +3,7 @@ mod config;
 mod consensus_protocol;
 mod era_supervisor;
 mod highway_core;
+mod metrics;
 mod protocols;
 #[cfg(test)]
 mod tests;
