+++ node/src/components/consensus/era_supervisor.rs
@@ -20,7 +20,7 @@ use blake2::{
 use datasize::DataSize;
 use fmt::Display;
 use num_traits::AsPrimitive;
-use prometheus::{Gauge, IntCounter, Registry};
+use prometheus::Registry;
 use rand::Rng;
 use serde::{Deserialize, Serialize};
 use tracing::{error, info, trace, warn};
@@ -41,6 +41,7 @@ use crate::{
                 BlockContext, ConsensusProtocol, ConsensusProtocolResult,
                 FinalizedBlock as CpFinalizedBlock,
             },
+            metrics::ConsensusMetrics,
             highway_core::{highway::Params, validators::Validators},
             protocols::highway::{HighwayContext, HighwayProtocol, HighwaySecret},
             traits::NodeIdT,
@@ -686,45 +687,5 @@ where
                     }
                 }),
         }
-    }
-}
-
-/// Network metrics to track Consensus
-#[derive(Debug)]
-pub struct ConsensusMetrics {
-    /// Gauge to track time between proposal and finalization.
-    finalization_time: Gauge,
-    /// Amount of finalized blocks.
-    finalized_block_count: IntCounter,
-    /// registry component.
-    registry: Registry,
-}
-
-impl ConsensusMetrics {
-    fn new(registry: &Registry) -> Result<Self, prometheus::Error> {
-        let finalization_time = Gauge::new(
-            "finalization_time",
-            "the amount of time, in milliseconds, between proposal and finalization of a block",
-        )?;
-        let finalized_block_count =
-            IntCounter::new("amount_of_blocks", "the number of blocks finalized so far")?;
-        registry.register(Box::new(finalization_time.clone()))?;
-        registry.register(Box::new(finalized_block_count.clone()))?;
-        Ok(ConsensusMetrics {
-            finalization_time,
-            finalized_block_count,
-            registry: registry.clone(),
-        })
-    }
-}
-
-impl Drop for ConsensusMetrics {
-    fn drop(&mut self) {
-        self.registry
-            .unregister(Box::new(self.finalization_time.clone()))
-            .expect("did not expect deregistering rate to fail");
-        self.registry
-            .unregister(Box::new(self.finalized_block_count.clone()))
-            .expect("did not expect deregisterting amount to fail");
     }
 }
